Introduce an action that is can optionally be called after the message ID for a message has been finalized but before it is actually sent.
Updated DiscoveryCustomerHandler to use the action to record the actual Message ID of any Discovery messages sent to a store.  Also updated the dictionary used to track the requests to be a concurrent dictionary to avoid any potential asynchronous update issues.

Removed NewMessageId from IEtpSessions as it should not be exposed externally.

Related Work Items: #27618